refactor: split gitlab.routes.ts into focused route sub-plugins (#321)
Split the 1058-line gitlab.routes.ts into 4 focused route sub-plugins:

- gitlab-integration.routes.ts: validate, integration CRUD, list repos
- gitlab-repository.routes.ts: link/unlink, branches, files listing
- gitlab-sync.routes.ts: export, import, operations, detect-conflicts
- gitlab-files.routes.ts: stored files GET, file content PUT

Extracted shared helpers (getAuthenticatedUserId, handleKnownRouteErrors)
into lib/gitlab-route-helpers.ts. Removed dead type-alias comments, empty
sections, and redundant content check covered by schema validation.

Reviewed by oracle. Behavior-preserving refactor; all existing tests pass.
